separate the service from the adapter
